fix: 颜色存储格式统一 - 彻底解决新项目颜色不显示
根因: 预设项目用 0xFF4CAF50 (Int→Long = -1090713120L)
      新建项目用 Color.value.toLong() (ULong→Long = 4283215696L)
      两者Long值不同，比较失败 → 所有颜色显示为未选中

修复: 统一使用 0xFF4CAF50L 或 .toInt().toLong() 格式
      确保存储和比较使用相同的Long值

+++ app/src/main/java/com/bulubulu/recordlifeitems/ui/projects/ProjectEditScreen.kt
@@ -164,7 +164,7 @@ fun ProjectEditScreen(
                 ProjectColors.forEach { color ->
                     val colorLong = color.value.toLong()
                     // Normalize: compare as Int to handle both -1090713120L and 4283215696L
-                    val isSelected = color.value.toLong() == selectedColor
+                    val isSelected = color.value.toInt().toLong() == selectedColor
                     Box(
                         modifier = Modifier
                             .size(40.dp)
@@ -177,7 +177,7 @@ fun ProjectEditScreen(
                                     Modifier
                                 }
                             )
-                            .clickable { viewModel.updateColor(color.value.toLong()) }
+                            .clickable { viewModel.updateColor(color.value.toInt().toLong()) }
                     ) {
                         if (isSelected) {
                             Box(
